+++ alicloud/data_source_alicloud_zones_test.go
@@ -5,6 +5,8 @@ import (
 	"strconv"
 	"testing"
 
+	"regexp"
+
 	"github.com/hashicorp/terraform/helper/resource"
 	"github.com/hashicorp/terraform/terraform"
 )
@@ -70,6 +72,24 @@ func TestAccAlicloudZonesDataSource_unitRegion(t *testing.T) {
 	})
 }
 
+func TestAccAlicloudZonesDataSource_multiZone(t *testing.T) {
+	resource.Test(t, resource.TestCase{
+		PreCheck: func() {
+			testAccPreCheck(t)
+		},
+		Providers: testAccProviders,
+		Steps: []resource.TestStep{
+			{
+				Config: testAccCheckAlicloudZonesDataSource_multiZone,
+				Check: resource.ComposeTestCheckFunc(
+					testAccCheckAlicloudDataSourceID("data.alicloud_zones.default"),
+					resource.TestMatchResourceAttr("data.alicloud_zones.default", "zones.0.id", regexp.MustCompile(fmt.Sprintf(".%s.", MULTI_IZ_SYMBOL))),
+				),
+			},
+		},
+	})
+}
+
 // the zone length changed occasionally
 // check by range to avoid test case failure
 func testCheckZoneLength(name string) resource.TestCheckFunc {
@@ -113,6 +133,10 @@ data "alicloud_zones" "foo" {
 `
 
 const testAccCheckAlicloudZonesDataSourceFilterIoOptimized = `
+provider "alicloud" {
+  region = "cn-shanghai"
+}
+
 data "alicloud_zones" "foo" {
 	available_instance_type= "ecs.c4.xlarge"
 	available_resource_creation= "IoOptimized"
@@ -131,3 +155,13 @@ data "alicloud_zones" "foo" {
 	available_resource_creation= "VSwitch"
 }
 `
+
+const testAccCheckAlicloudZonesDataSource_multiZone = `
+provider "alicloud" {
+  region = "cn-shanghai"
+}
+
+data "alicloud_zones" "default" {
+  available_resource_creation= "Rds"
+  multi = true
+}`